Source code structure refactoring
