+++ TDE/GCC_class.cpp
