+++ TDE/GPS_class.cpp
