--- TDE_TOOLS/core/CMakeLists.txt
@@ -1,5 +1,0 @@
-project(CORE)
-
-cmake_minimum_required(VERSION 3.16)
-
-add_library(core STATIC ${CMAKE_SOURCE_DIR}/${CMAKE_PROJECT_NAME}/core/core.cpp)
