+++ cmd_util/CMakeLists.txt
@@ -1,0 +1,8 @@
+project(cmd_util)
+
+cmake_minimum_required(VERSION 3.16)
+
+add_executable(${PROJECT_NAME}
+            ${CMAKE_CURRENT_SOURCE_DIR}/main.cpp
+            ${CMAKE_CURRENT_SOURCE_DIR}/program_environment.cpp)
+target_link_libraries(${PROJECT_NAME} TDE_calculator fft)
