+++ cmd_util/main.cpp
@@ -5,7 +5,7 @@
 #include "AudioFile.h"
 #include "main.hpp"
 #include "TDE_class.hpp"
-#include "program_environment/program_environment.hpp"
+#include <program_environment.hpp>
 
 static const tde_meth DEFAULT_TDE_METHOD = GPS_TDE;
 static const uint16_t DEFAULT_SAMPLE_RATE = 44100;
