+++ fft/fft_forward_class.cpp
