+++ internal_include/core.hpp
